Update Get-HostPoolInfo.ps1
Modified the script so that personal machines that are unavailable (because hibernation) or shutdown (because scaling plan) are ignored. Also fixed this issue:
The issue is that $HostList is being reused in a loop without resetting it, causing PowerShell to attempt adding properties that already exist.

Instead of a single $HostList object, create an array of objects to store multiple session hosts.